+++ pseudocode.py
@@ -1,0 +1,5 @@
+# There are 2 different inventories including healables and tools
+# The healables will be: Knife, Flaregun, Metal
+# The tools will be: Bandages, Medkit, Painkiller
+
+# The character will be able to move forward, back, right and left