+++ changelog.md
@@ -1,4 +1,6 @@
 # RPG: Survival Instinct Changelog
 v0.0
 -added READ.me
--added Changelog
+-added changelog
+-added main game file
+-added pseudocode file