+++ changelog.md
@@ -2,5 +2,5 @@
 v0.0
 -added READ.me
 -added changelog
--added main game file
--added pseudocode file
+-added main game file with RPG Numerical lists
+-added pseudocode file with comments