+++ changelog.md
@@ -2,5 +2,5 @@
 v0.0
 - added READ.me for important information
 - added changelog to show modifications
-- added main game file with RPG Numerical lists
+- added main game file with RPG Nested Dictionaries
 - added pseudocode file with comments