+++ changelog.md
@@ -4,3 +4,5 @@ v0.0
 - added changelog to show modifications
 - added main game file with RPG Nested Dictionaries
 - added pseudocode file with comments
+- added menu file with functions
+- added map file for map game